I have the rectangle box moving now in response to some GLUT keyboard input, and I disabled the camera controls, which use the same WASD and space keys.  I can now add immediate force vectors to influence the XZ plane motion of the box.  Next: reflect linear velocity around a vector (this is crude 'bounce off wall' behavior).

+++ Sandbox_Game_2/Engine_2/Timing/Game_Clock.cpp
@@ -34,7 +34,7 @@ namespace Timing
       m_delta_time_last_frame = m_stopwatch.lap();
    }
 
-   float Game_Clock::get_delta_time_last_frame() const
+   double Game_Clock::get_delta_time_last_frame() const
    {
       return m_delta_time_last_frame;
    }
